server/requirements/develop.txt
# Development dependencies
-r _base.txt

django-debug-toolbar==1.9.1
